undo strict period handling
The control loop (that should run at 50Hz = 20ms) is often much slower than 20ms.
In a short test, we have seen a lag of up to 4ms. Trying to adhere to the 20ms period
seems to end up in a situation, where the main humotion control loop runs without
any delay. In this case, the real (not the virtual) bus seems to get screwed.